+++ pyleco/test.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 from typing import Any, Optional, Sequence, Union
 
 from .core.message import Message