+++ pyleco/utils/coordinator_utils.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 from abc import abstractmethod
 from dataclasses import dataclass
 import logging