+++ pyleco/utils/extended_message_handler.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 import json
 import pickle
 from typing import Optional