+++ tests/coordinators/test_coordinator.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 from unittest.mock import MagicMock
 
 import pytest