+++ tests/utils/test_base_communicator.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 import logging
 import time
 from typing import Optional